♻️ Refactor utils into a separate module
