+++ repid/queue.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
+from repid._utils import VALID_NAME
 from repid.main import Repid
-from repid.utils import VALID_NAME
 
 if TYPE_CHECKING:
     from repid.connection import Connection
